requirements/dev.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements/dev.in
#
asgiref==3.5.2
    # via django
attrs==22.1.0
    # via
    #   jsonschema
    #   pytest
django==4.1.3
    # via
    #   -r requirements/base.in
    #   django-filter
    #   djangorestframework
    #   drf-spectacular
django-filter==22.1
    # via -r requirements/base.in
djangorestframework==3.14.0
    # via
    #   -r requirements/base.in
    #   drf-spectacular
drf-spectacular==0.24.2
    # via -r requirements/base.in
exceptiongroup==1.0.1
    # via pytest
factory-boy==3.2.1
    # via -r requirements/base.in
faker==15.3.4
    # via factory-boy
inflection==0.5.1
    # via drf-spectacular
iniconfig==1.1.1
    # via pytest
isort==5.10.1
    # via -r requirements/dev.in
jsonschema==4.17.0
    # via drf-spectacular
packaging==21.3
    # via pytest
pluggy==1.0.0
    # via pytest
psycopg2==2.9.5
    # via -r requirements/base.in
pyparsing==3.0.9
    # via packaging
pyrsistent==0.19.2
    # via jsonschema
pytest==7.2.0
    # via -r requirements/base.in
python-dateutil==2.8.2
    # via faker
pytz==2022.6
    # via djangorestframework
pyyaml==6.0
    # via drf-spectacular
six==1.16.0
    # via python-dateutil
sqlparse==0.4.3
    # via django
tomli==2.0.1
    # via pytest
uritemplate==4.1.1
    # via drf-spectacular
